twitter client refactor
