+++ twitter/users.csv
@@ -1,10 +1,20 @@
-budougumi0617
-joncalhoun
-botgopher
-bhfaught
-DabiddoKun
-jlucktay
-conradwt
-cool_golang
-nytro04
-nand0tech
+33Zilda
+LuluIraus
+FortesGil
+NoceMassao
+Ramonabrt
+Fernand63139206
+Federic77565028
+jorgenilsonma
+EvaCarvalho
+FariaAyrton
+agentebaiano007
+JoceyrR
+magagonzaless
+aquelado1
+Carlos35495410
+nilsonnoslinho1
+Sd6a
+Ivete58505648
+ruyganeff
+remodest
